+++ test/fuzz/NodeFuzz.t.sol
@@ -254,12 +254,12 @@ contract NodeFuzzTest is BaseTest {
                         FEE PAYMENTS
     //////////////////////////////////////////////////////////////*/
 
-    function test_fuzz_node_payManagementFees(uint64 annualFee, uint256 protocolFee, uint256 seedAmount) public {
+    function test_fuzz_node_payManagementFees(uint64 annualFee, uint64 protocolFee, uint256 seedAmount) public {
         address ownerFeesRecipient = makeAddr("ownerFeesRecipient");
         address protocolFeesRecipient = makeAddr("protocolFeesRecipient");
 
         annualFee = uint64(bound(annualFee, 0, 1e18));
-        protocolFee = bound(protocolFee, 0, 1e18);
+        protocolFee = uint64(bound(protocolFee, 0, 1e18));
         seedAmount = bound(seedAmount, 1e18, 1e36);
 
         vm.startPrank(owner);
@@ -287,7 +287,7 @@ contract NodeFuzzTest is BaseTest {
 
     function test_fuzz_node_payManagementFees_different_durations(
         uint64 annualFee,
-        uint256 protocolFee,
+        uint64 protocolFee,
         uint256 seedAmount,
         uint256 duration
     ) public {
@@ -295,7 +295,7 @@ contract NodeFuzzTest is BaseTest {
         address protocolFeesRecipient = makeAddr("protocolFeesRecipient");
 
         annualFee = uint64(bound(annualFee, 0, 1e18));
-        protocolFee = bound(protocolFee, 0, 1e18);
+        protocolFee = uint64(bound(protocolFee, 0, 1e18));
         seedAmount = bound(seedAmount, 1e18, 1e36);
         duration = bound(duration, 1 days, 365 days);
