+++ test/unit/routers/ERC7540Router.t.sol
@@ -701,7 +701,7 @@ contract ERC7540RouterTest is BaseTest {
         router7540.setWhitelistStatus(address(liquidityPool), true);
         vm.stopPrank();
 
-        uint256 expectedDeposit = 100 ether * node.getComponentRatio(address(liquidityPool)) / 1 ether;
+        uint256 expectedDeposit = 100 ether * uint256(node.getComponentRatio(address(liquidityPool))) / 1 ether;
 
         vm.startPrank(rebalancer);
         node.startRebalance();
